+++ user-service/src/main/java/ru/clevertec/user/UserServiceApplication.java
@@ -1,4 +1,4 @@
-package ru.clevertec.users;
+package ru.clevertec.user;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
