+++ user-service/src/test/java/ru/clevertec/user/UserServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.clevertec.users;
+package ru.clevertec.user;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
